requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2

# 安全和开发依赖
python-dotenv==1.0.0
email-validator==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# 异步数据库支持
aiosqlite==0.19.0
databases[sqlite]==0.8.0
torch==2.1.1
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
qrcode==7.4.2
ccxt==4.1.64
asyncio-mqtt==0.16.1
influxdb-client==1.38.0
celery==5.3.4
flower==2.0.1
gunicorn==21.2.0
aiohttp==3.9.1
